# homeworks/Day3/sessionworks/docs/api.yml
openapi: 3.0.3
info:
  title: Mini Attack Surface Management API
  description: |
    API cho hệ thống quản lý Attack Surface Management (EASM).
    Hỗ trợ quản lý assets và thực hiện các loại scan bảo mật.
  version: 2.0.0
  contact:
    name: Intern Program

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Kiểm tra trạng thái service và database connection
      operationId: healthCheck
      tags:
        - System
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  database:
                    type: string
                    example: connected

  /assets:
    post:
      summary: Create a new asset
      description: Tạo mới một asset (domain, IP, hoặc service)
      operationId: createAsset
      tags:
        - Assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: example.com
                type:
                  type: string
                  enum: [domain, ip, service]
                  example: domain
                status:
                  type: string
                  enum: [active, inactive]
                  default: active
      responses:
        "201":
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all assets
      description: Lấy danh sách assets với filter, search, pagination
      operationId: listAssets
      tags:
        - Assets
      parameters:
        - name: type
          in: query
          description: Filter theo loại asset
          schema:
            type: string
            enum: [domain, ip, service]
        - name: status
          in: query
          description: Filter theo trạng thái
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          description: Tìm kiếm theo tên (partial match)
          schema:
            type: string
        - name: page
          in: query
          description: Số trang
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Số items mỗi trang
          schema:
            type: integer
            default: 20
        - name: sort_by
          in: query
          description: Sắp xếp theo field
          schema:
            type: string
            enum: [name, type, status, created_at]
            default: created_at
        - name: sort_order
          in: query
          description: Thứ tự sắp xếp
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAssets"

  /assets/{id}:
    get:
      summary: Get a single asset
      operationId: getAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Asset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update an asset
      operationId: updateAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                type:
                  type: string
                  enum: [domain, ip, service]
                status:
                  type: string
                  enum: [active, inactive]
      responses:
        "200":
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete an asset
      operationId: deleteAsset
      tags:
        - Assets
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset deleted successfully
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}/scan:
    post:
      summary: Start a scan job
      description: Bắt đầu scan cho asset. Trả về 202 ngay, scan chạy async.
      operationId: startScan
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scan_type
              properties:
                scan_type:
                  type: string
                  enum: [dns, whois, subdomain, ip, port, ssl, tech]
                  example: dns
      responses:
        "202":
          description: Scan job created (async)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "400":
          description: Invalid scan type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}/scans:
    get:
      summary: List all scan jobs for an asset
      operationId: listAssetScans
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: List of scan jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanJob"

  /assets/{id}/results:
    get:
      summary: Get ALL scan results for an asset (aggregate)
      description: Trả về tất cả kết quả scan của asset, gộp theo loại scan.
      operationId: getAssetResults
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Aggregated scan results
          content:
            application/json:
              schema:
                type: object
                properties:
                  subdomains:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subdomain"
                  dns_records:
                    type: array
                    items:
                      $ref: "#/components/schemas/DNSRecord"
                  whois:
                    $ref: "#/components/schemas/WHOISRecord"
                  ip_scan:
                    type: array
                    items:
                      $ref: "#/components/schemas/IPScanResult"
                  port_scan:
                    type: array
                    items:
                      $ref: "#/components/schemas/PortScanResult"
                  ssl_scan:
                    type: array
                    items:
                      $ref: "#/components/schemas/SSLScanResult"
                  tech_scan:
                    type: array
                    items:
                      $ref: "#/components/schemas/TechScanResult"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}/subdomains:
    get:
      summary: Get subdomains for an asset
      operationId: getAssetSubdomains
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: List of subdomains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subdomain"

  /assets/{id}/dns:
    get:
      summary: Get DNS records for an asset
      operationId: getAssetDNS
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: List of DNS records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DNSRecord"

  /assets/{id}/whois:
    get:
      summary: Get WHOIS info for an asset
      operationId: getAssetWHOIS
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: WHOIS record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WHOISRecord"

  /scan-jobs/{id}:
    get:
      summary: Get scan job status
      operationId: getScanJob
      tags:
        - Scanning
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của scan job
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scan job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "404":
          description: Scan job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /scan-jobs/{id}/results:
    get:
      summary: Get scan job results
      description: Kết quả trả về phụ thuộc vào scan_type của job.
      operationId: getScanResults
      tags:
        - Scanning
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của scan job
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scan results (format depends on scan_type)
        "404":
          description: Scan job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      description: UUID của asset
      schema:
        type: string
        format: uuid

  schemas:
    Asset:
      type: object
      required:
        - id
        - name
        - type
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "example.com"
        type:
          type: string
          enum: [domain, ip, service]
          example: "domain"
        status:
          type: string
          enum: [active, inactive]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PaginatedAssets:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 3

    ScanJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_type:
          type: string
          enum: [dns, whois, subdomain, ip, port, ssl, tech]
        status:
          type: string
          enum: [pending, running, completed, failed, partial]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
        results:
          type: integer
        created_at:
          type: string
          format: date-time

    Subdomain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        name:
          type: string
          example: "api.example.com"
        source:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    DNSRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        record_type:
          type: string
          enum: [A, AAAA, CNAME, MX, NS, TXT, SOA]
        name:
          type: string
        value:
          type: string
        ttl:
          type: integer
        created_at:
          type: string
          format: date-time

    WHOISRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        registrar:
          type: string
        created_date:
          type: string
          format: date-time
          nullable: true
        expiry_date:
          type: string
          format: date-time
          nullable: true
        name_servers:
          type: string
          description: JSON array of nameservers
        status:
          type: string
        emails:
          type: string
          description: JSON array of contact emails
        raw_data:
          type: string
        created_at:
          type: string
          format: date-time

    IPScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        ip_address:
          type: string
          example: "104.21.32.1"
        geolocation:
          type: object
          properties:
            country:
              type: string
              example: "US"
            country_code:
              type: string
              example: "US"
            city:
              type: string
              example: "San Francisco"
            region:
              type: string
              example: "California"
            latitude:
              type: number
              example: 37.7749
            longitude:
              type: number
              example: -122.4194
            isp:
              type: string
            org:
              type: string
        asn:
          type: object
          properties:
            number:
              type: integer
              example: 13335
            name:
              type: string
              example: "CLOUDFLARENET"
            description:
              type: string
        reverse_dns:
          type: string
          example: "one.one.one.one"
        created_at:
          type: string
          format: date-time

    PortScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        ip_address:
          type: string
          example: "127.0.0.1"
        open_ports:
          type: array
          items:
            type: object
            properties:
              port:
                type: integer
                example: 22
              protocol:
                type: string
                example: "tcp"
              state:
                type: string
                example: "open"
              service:
                type: string
                example: "ssh"
              version:
                type: string
        closed_ports:
          type: integer
          example: 996
        total_scanned:
          type: integer
          example: 1000
        scan_duration_ms:
          type: integer
          example: 5432
        created_at:
          type: string
          format: date-time

    SSLScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        domain:
          type: string
          example: "example.com"
        certificate:
          type: object
          properties:
            subject:
              type: string
              example: "CN=example.com"
            issuer:
              type: string
              example: "CN=Let's Encrypt Authority X3"
            serial_number:
              type: string
            valid_from:
              type: string
              format: date-time
            valid_until:
              type: string
              format: date-time
            days_until_expiry:
              type: integer
              example: 45
            is_expired:
              type: boolean
            is_self_signed:
              type: boolean
            san:
              type: array
              items:
                type: string
        connection:
          type: object
          properties:
            tls_version:
              type: string
              example: "TLS 1.3"
            cipher_suite:
              type: string
              example: "TLS_AES_256_GCM_SHA384"
        grade:
          type: string
          enum: [A, B, C, F]
          example: "A"
        issues:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    TechScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        domain:
          type: string
          example: "example.com"
        technologies:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "nginx"
              category:
                type: string
                example: "Web Server"
              version:
                type: string
                example: "1.18.0"
                nullable: true
              confidence:
                type: integer
                example: 100
        headers:
          type: object
          additionalProperties:
            type: string
          example:
            server: "nginx/1.18.0"
            x-powered-by: "Express"
        meta_tags:
          type: object
          additionalProperties:
            type: string
          example:
            generator: "React 18.2.0"
        created_at:
          type: string
          format: date-time

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid input"